requirements.txt
fastapi
pydantic
uvicorn
typing
requests
pytest
httpx
python-dotenv

